+++ packages/riverpod/test/src/providers/stream_notifier_test.dart
@@ -8,73 +8,11 @@ import 'package:riverpod/riverpod.dart' hide ErrorListener;
 import 'package:test/test.dart';
 
 import '../../utils.dart';
-import 'factory.dart';
+import '../matrix/stream_notifier_provider.dart';
 
 void main() {
   for (final factory in matrix()) {
     group(factory.label, () {
-      test('Can assign `AsyncLoading<T>` to `AsyncValue<void>`', () {
-        // Regression test for https://github.com/rrousselGit/riverpod/issues/2120
-        final provider = factory.simpleTestProvider<void>((ref) => 42);
-        final container = createContainer();
-
-        final sub = container.listen(provider.notifier, (prev, next) {});
-
-        // ignore: void_checks
-        expect(sub.read().state, const AsyncData<void>(42));
-
-        sub.read().state = const AsyncLoading<int>();
-
-        expect(
-          sub.read().state,
-          isA<AsyncLoading<void>>()
-              .having((e) => e.hasValue, 'hasValue', true)
-              .having((e) => e.value, 'value', 42),
-        );
-      });
-
-      test('Can assign `AsyncData<T>` to `AsyncValue<void>`', () {
-        // Regression test for https://github.com/rrousselGit/riverpod/issues/2120
-        final provider = factory.simpleTestProvider<void>((ref) => 42);
-        final container = createContainer();
-
-        final sub = container.listen(provider.notifier, (prev, next) {});
-
-        // ignore: void_checks
-        expect(sub.read().state, const AsyncData<void>(42));
-
-        sub.read().state = const AsyncData<int>(42);
-
-        expect(
-          sub.read().state,
-          isA<AsyncData<void>>()
-              .having((e) => e.hasValue, 'hasValue', true)
-              .having((e) => e.value, 'value', 42),
-        );
-      });
-
-      test('Can assign `AsyncError<T>` to `AsyncValue<void>`', () {
-        // Regression test for https://github.com/rrousselGit/riverpod/issues/2120
-        final provider = factory.simpleTestProvider<void>((ref) => 42);
-        final container = createContainer();
-
-        final sub = container.listen(provider.notifier, (prev, next) {});
-
-        // ignore: void_checks
-        expect(sub.read().state, const AsyncData<void>(42));
-
-        sub.read().state = AsyncError<int>(21, StackTrace.current);
-
-        expect(
-          sub.read().state,
-          isA<AsyncError<void>>()
-              .having((e) => e.hasValue, 'hasValue', true)
-              .having((e) => e.value, 'value', 42)
-              .having((e) => e.hasError, 'hasError', true)
-              .having((e) => e.error, 'error', 21),
-        );
-      });
-
       group('supports AsyncValue transition', () {
         test(
             'performs seamless copyWithPrevious if triggered by ref.invalidate/ref.refresh',
@@ -82,7 +20,7 @@ void main() {
           final container = createContainer();
           var count = 0;
           final provider = factory.simpleTestProvider(
-            (ref) => Future.value(count++),
+            (ref) => Stream.value(count++),
           );
 
           container.listen(provider, (previous, next) {});
@@ -112,7 +50,7 @@ void main() {
             'performs seamless:false copyWithPrevious on `state = AsyncLoading()`',
             () async {
           final container = createContainer();
-          final provider = factory.simpleTestProvider((ref) => Future.value(0));
+          final provider = factory.simpleTestProvider((ref) => Stream.value(0));
 
           final sub = container.listen(provider.notifier, (previous, next) {});
 
@@ -134,7 +72,7 @@ void main() {
           final container = createContainer();
           final dep = StateProvider((ref) => 0);
           final provider = factory.simpleTestProvider(
-            (ref) => Future.value(ref.watch(dep)),
+            (ref) => Stream.value(ref.watch(dep)),
           );
 
           container.listen(provider, (previous, next) {});
@@ -153,39 +91,13 @@ void main() {
           expect(container.read(provider), const AsyncData(1));
         });
 
-        test('performs seamless data > loading > error transition', () async {
-          final container = createContainer();
-          var result = Future.value(42);
-          final provider = FutureProvider((ref) => result);
-
-          final sub = container.listen(provider.future, (_, __) {});
-
-          expect(container.read(provider), const AsyncLoading<int>());
-          expect(await sub.read(), 42);
-          expect(container.read(provider), const AsyncData<int>(42));
-
-          result = Future.error('err', StackTrace.empty);
-          container.invalidate(provider);
-
-          expect(
-            container.read(provider),
-            const AsyncLoading<int>().copyWithPrevious(const AsyncData(42)),
-          );
-          await expectLater(sub.read(), throwsA('err'));
-          expect(
-            container.read(provider),
-            const AsyncError<int>('err', StackTrace.empty)
-                .copyWithPrevious(const AsyncData<int>(42)),
-          );
-        });
-
         test(
             'performs seamless:false copyWithPrevious if both triggered by a dependency change and ref.refresh',
             () async {
           final container = createContainer();
           final dep = StateProvider((ref) => 0);
           final provider = factory.simpleTestProvider(
-            (ref) => Future.value(ref.watch(dep)),
+            (ref) => Stream.value(ref.watch(dep)),
           );
 
           container.listen(provider, (previous, next) {});
@@ -206,7 +118,7 @@ void main() {
       });
 
       test('does not notify listeners when refreshed during loading', () async {
-        final provider = factory.simpleTestProvider((ref) => Future.value(0));
+        final provider = factory.simpleTestProvider((ref) => Stream.value(0));
         final container = createContainer();
         final listener = Listener<AsyncValue<int>>();
 
@@ -254,9 +166,9 @@ void main() {
       });
 
       test(
-          'converts AsyncNotifier.build into an AsyncData if the future completes',
+          'converts StreamNotifier.build into an AsyncData if the future completes',
           () async {
-        final provider = factory.simpleTestProvider((ref) => Future.value(0));
+        final provider = factory.simpleTestProvider((ref) => Stream.value(0));
         final container = createContainer();
         final listener = Listener<AsyncValue<int>>();
 
@@ -281,10 +193,10 @@ void main() {
       });
 
       test(
-          'converts AsyncNotifier.build into an AsyncError if the future fails',
+          'converts StreamNotifier.build into an AsyncError if the future fails',
           () async {
         final provider = factory.simpleTestProvider<int>(
-          (ref) => Future.error(0, StackTrace.empty),
+          (ref) => Stream.error(0, StackTrace.empty),
         );
         final container = createContainer();
         final listener = Listener<AsyncValue<int>>();
@@ -309,7 +221,7 @@ void main() {
         );
       });
 
-      test('supports cases where the AsyncNotifier constructor throws',
+      test('supports cases where the StreamNotifier constructor throws',
           () async {
         final provider = factory.testProvider<int>(
           () => Error.throwWithStackTrace(0, StackTrace.empty),
@@ -332,21 +244,7 @@ void main() {
       });
 
       test(
-          'synchronously emits AsyncData if AsyncNotifier.build emits synchronously',
-          () async {
-        final provider = factory.simpleTestProvider<int>((ref) => 0);
-        final container = createContainer();
-        final listener = Listener<AsyncValue<int>>();
-
-        container.listen(provider, listener.call, fireImmediately: true);
-
-        verifyOnly(listener, listener(null, const AsyncData(0)));
-        expect(container.read(provider.notifier).state, const AsyncData(0));
-        await expectLater(container.read(provider.future), completion(0));
-      });
-
-      test(
-          'synchronously emits AsyncError if AsyncNotifier.build throws synchronously',
+          'synchronously emits AsyncError if StreamNotifier.build throws synchronously',
           () async {
         final provider = factory.simpleTestProvider<int>(
           (ref) => Error.throwWithStackTrace(42, StackTrace.empty),
@@ -377,7 +275,7 @@ void main() {
           1: Completer<int>.sync(),
         };
         final provider = factory.simpleTestProvider<int>(
-          (ref) => completers[ref.watch(dep)]!.future,
+          (ref) => Stream.fromFuture(completers[ref.watch(dep)]!.future),
         );
         final listener = Listener<AsyncValue<int>>();
 
@@ -416,7 +314,7 @@ void main() {
           1: Completer<int>.sync(),
         };
         final provider = factory.simpleTestProvider<int>(
-          (ref) => completers[ref.watch(dep)]!.future,
+          (ref) => Stream.fromFuture(completers[ref.watch(dep)]!.future),
         );
         final listener = Listener<AsyncValue<int>>();
 
@@ -448,14 +346,17 @@ void main() {
         );
       });
 
-      group('AsyncNotifier.state', () {
+      group('StreamNotifier.state', () {
         test(
             'when manually modifying the state, the new exposed value contains the previous state when possible',
             () async {
-          final provider = factory.simpleTestProvider<int>((ref) => 0);
+          final provider = factory.simpleTestProvider<int>(
+            (ref) => Stream.value(0),
+          );
           final container = createContainer();
 
           final sub = container.listen(provider.notifier, (previous, next) {});
+          await container.read(provider.future);
 
           // ignore: prefer_const_constructors, not using `const` as we voluntarility break identity to test `identical`
           final newState = AsyncData(84);
@@ -495,7 +396,7 @@ void main() {
           final provider = factory.simpleTestProvider<int>(
             (ref) {
               listener();
-              return Future.value(ref.watch(dep));
+              return Stream.value(ref.watch(dep));
             },
           );
           final container = createContainer();
@@ -525,11 +426,11 @@ void main() {
           late AsyncValue<int> state;
           final provider = factory.testProvider<int>(
             () {
-              late AsyncTestNotifierBase<int> notifier;
+              late StreamTestNotifierBase<int> notifier;
               return notifier = factory.notifier<int>(
                 (ref) {
                   state = notifier.state;
-                  return Future.value(ref.watch(dep));
+                  return Stream.value(ref.watch(dep));
                 },
               );
             },
@@ -549,10 +450,13 @@ void main() {
           );
         });
 
-        test('notifies listeners when the setter is called', () {
-          final provider = factory.simpleTestProvider((ref) => 0);
+        test('notifies listeners when the setter is called', () async {
+          final provider = factory.simpleTestProvider((ref) => Stream.value(0));
           final container = createContainer();
           final listener = Listener<AsyncValue<int>>();
+
+          // Skip the loading
+          await container.listen(provider.future, (previous, next) {}).read();
 
           container.listen(provider, listener.call);
 
@@ -567,14 +471,14 @@ void main() {
         });
       });
 
-      group('AsyncNotifier.future', () {
+      group('StreamNotifier.future', () {
         test(
-            'when disposed during loading, resolves with the content of AsyncNotifier.build',
+            'when disposed during loading, resolves with the content of StreamNotifier.build',
             () async {
           final container = createContainer();
           final completer = Completer<int>.sync();
           final provider = factory.simpleTestProvider<int>(
-            (ref) => completer.future,
+            (ref) => Stream.fromFuture(completer.future),
           );
 
           final future = container.read(provider.future);
@@ -586,12 +490,12 @@ void main() {
         });
 
         test(
-            'when disposed during loading, resolves with the error of AsyncNotifier.build',
+            'when disposed during loading, resolves with the error of StreamNotifier.build',
             () async {
           final container = createContainer();
           final completer = Completer<int>.sync();
           final provider = factory.simpleTestProvider<int>(
-            (ref) => completer.future,
+            (ref) => Stream.fromFuture(completer.future),
           );
 
           final future = container.read(provider.future);
@@ -610,7 +514,7 @@ void main() {
             final container = createContainer();
             final completer = Completer<int>.sync();
             final provider = factory.simpleTestProvider<int>(
-              (ref) => completer.future,
+              (ref) => Stream.fromFuture(completer.future),
             );
 
             container.read(provider);
@@ -633,7 +537,7 @@ void main() {
             final container = createContainer();
             final completer = Completer<int>.sync();
             final provider = factory.simpleTestProvider<int>(
-              (ref) => completer.future,
+              (ref) => Stream.fromFuture(completer.future),
             );
 
             container.read(provider);
@@ -652,12 +556,12 @@ void main() {
         );
 
         test(
-            'resolves with the new state if AsyncNotifier.state is modified during loading',
+            'resolves with the new state if StreamNotifier.state is modified during loading',
             () async {
           final container = createContainer();
           final completer = Completer<int>.sync();
           final provider = factory.simpleTestProvider<int>(
-            (ref) => completer.future,
+            (ref) => Stream.fromFuture(completer.future),
           );
           final listener = Listener<Future<int>>();
 
@@ -682,7 +586,9 @@ void main() {
         test('resolves with the new state when notifier.state is changed',
             () async {
           final container = createContainer();
-          final provider = factory.simpleTestProvider<int>((ref) => 0);
+          final provider = factory.simpleTestProvider<int>(
+            (ref) => Stream.value(0),
+          );
           final listener = Listener<Future<int>>();
 
           final sub = container.listen(provider.notifier, (previous, next) {});
@@ -709,7 +615,7 @@ void main() {
           final provider = factory.simpleTestProvider<int>(
             (ref) {
               listener();
-              return Future.value(ref.watch(dep));
+              return Stream.value(ref.watch(dep));
             },
           );
           final container = createContainer();
@@ -728,7 +634,7 @@ void main() {
           final provider = factory.simpleTestProvider<int>(
             (ref) {
               listener();
-              return Future.value(ref.watch(dep));
+              return Stream.value(ref.watch(dep));
             },
           );
           final container = createContainer();
@@ -748,7 +654,7 @@ void main() {
         });
       });
 
-      group('AsyncNotifierProvider.notifier', () {
+      group('StreamNotifierProvider.notifier', () {
         test(
             'never emits an update. The Notifier is never recreated once it is instantiated',
             () async {
@@ -756,9 +662,14 @@ void main() {
           final dep = StateProvider((ref) => 0);
           final provider = factory.testProvider<int>(() {
             listener();
-            return factory.notifier((ref) => ref.watch(dep));
+            return factory.notifier(
+              (ref) => Stream.value(ref.watch(dep)),
+            );
           });
           final container = createContainer();
+
+          // Skip the loading
+          await container.listen(provider.future, (previous, next) {}).read();
 
           container.listen(provider, (previous, next) {});
           final notifier = container.read(provider.notifier);
@@ -767,6 +678,7 @@ void main() {
           expect(container.read(provider), const AsyncData(0));
 
           container.read(dep.notifier).state++;
+          await container.read(provider.future);
 
           expect(container.read(provider), const AsyncData(1));
           expect(container.read(provider.notifier), same(notifier));
@@ -775,14 +687,17 @@ void main() {
       });
 
       test(
-          'Can override AsyncNotifier.updateShouldNotify to change the default filter logic',
-          () {
+          'Can override StreamNotifier.updateShouldNotify to change the default filter logic',
+          () async {
         final provider = factory.simpleTestProvider<Equal<int>>(
-          (ref) => Equal(42),
+          (ref) => Stream.value(Equal(42)),
           updateShouldNotify: (a, b) => a != b,
         );
         final container = createContainer();
         final listener = Listener<AsyncValue<Equal<int>>>();
+
+        // Skip the loading
+        await container.listen(provider.future, (previous, next) {}).read();
 
         container.listen(provider, listener.call);
         final notifier = container.read(provider.notifier);
@@ -809,8 +724,11 @@ void main() {
         test('passes in the latest state', () async {
           final container = createContainer();
           final provider = factory.simpleTestProvider<int>(
-            (ref) => 0,
+            (ref) => Stream.value(0),
           );
+
+          // Skip the loading
+          await container.listen(provider.future, (previous, next) {}).read();
 
           final sub = container.listen(provider.notifier, (prev, next) {});
 
@@ -873,7 +791,12 @@ void main() {
             'executes immediately with current state if a state is avalailable',
             () async {
           final container = createContainer();
-          final provider = factory.simpleTestProvider<int>((ref) => 1);
+          final provider = factory.simpleTestProvider<int>(
+            (ref) => Stream.value(1),
+          );
+
+          // Skip the loading
+          await container.listen(provider.future, (previous, next) {}).read();
 
           final sub = container.listen(provider.notifier, (prev, next) {});
 
@@ -920,7 +843,7 @@ void main() {
         test('awaits the future resolution if in loading state', () async {
           final container = createContainer();
           final provider = factory.simpleTestProvider<int>(
-            (ref) => Future.value(42),
+            (ref) => Stream.value(42),
           );
 
           final sub = container.listen(provider.notifier, (prev, next) {});
@@ -937,61 +860,78 @@ void main() {
     });
   }
 
-  test('supports overrideWith', () {
-    final provider = AsyncNotifierProvider<AsyncTestNotifier<int>, int>(
-      () => AsyncTestNotifier((ref) => 0),
+  test('supports overrideWith', () async {
+    final provider = StreamNotifierProvider<StreamTestNotifier<int>, int>(
+      () => StreamTestNotifier((ref) => Stream.value(0)),
     );
-    final autoDispose = AsyncNotifierProvider.autoDispose<
-        AutoDisposeAsyncTestNotifier<int>, int>(
-      () => AutoDisposeAsyncTestNotifier((ref) => 0),
+    final autoDispose = StreamNotifierProvider.autoDispose<
+        AutoDisposeStreamTestNotifier<int>, int>(
+      () => AutoDisposeStreamTestNotifier((ref) => Stream.value(0)),
     );
     final container = createContainer(
       overrides: [
-        provider.overrideWith(() => AsyncTestNotifier((ref) => 42)),
+        provider.overrideWith(
+          () => StreamTestNotifier((ref) => Stream.value(42)),
+        ),
         autoDispose.overrideWith(
-          () => AutoDisposeAsyncTestNotifier((ref) => 84),
+          () => AutoDisposeStreamTestNotifier((ref) => Stream.value(84)),
         ),
       ],
     );
+
+    // Skip the loading
+    await container.listen(provider.future, (previous, next) {}).read();
+    await container.listen(autoDispose.future, (previous, next) {}).read();
 
     expect(container.read(provider).value, 42);
     expect(container.read(autoDispose).value, 84);
   });
 
-  test('supports family overrideWith', () {
+  test('supports family overrideWith', () async {
     final family =
-        AsyncNotifierProvider.family<AsyncTestNotifierFamily<int>, int, int>(
-      () => AsyncTestNotifierFamily<int>((ref) => 0),
+        StreamNotifierProvider.family<StreamTestNotifierFamily<int>, int, int>(
+      () => StreamTestNotifierFamily<int>((ref) => Stream.value(0)),
     );
-    final autoDisposeFamily = AsyncNotifierProvider.autoDispose
-        .family<AutoDisposeAsyncTestNotifierFamily<int>, int, int>(
-      () => AutoDisposeAsyncTestNotifierFamily<int>((ref) => 0),
+    final autoDisposeFamily = StreamNotifierProvider.autoDispose
+        .family<AutoDisposeStreamTestNotifierFamily<int>, int, int>(
+      () => AutoDisposeStreamTestNotifierFamily<int>((ref) => Stream.value(0)),
     );
     final container = createContainer(
       overrides: [
         family.overrideWith(
-          () => AsyncTestNotifierFamily<int>((ref) => 42),
+          () => StreamTestNotifierFamily<int>((ref) => Stream.value(42)),
         ),
         autoDisposeFamily.overrideWith(
-          () => AutoDisposeAsyncTestNotifierFamily<int>((ref) => 84),
+          () => AutoDisposeStreamTestNotifierFamily<int>(
+            (ref) => Stream.value(84),
+          ),
         ),
       ],
     );
+
+    // Skip the loading
+    await container.listen(family(10).future, (previous, next) {}).read();
+    await container
+        .listen(autoDisposeFamily(10).future, (previous, next) {})
+        .read();
 
     expect(container.read(family(10)).value, 42);
     expect(container.read(autoDisposeFamily(10)).value, 84);
   });
 
   group('AutoDispose variant', () {
-    test('can watch autoDispose providers', () {
+    test('can watch autoDispose providers', () async {
       final dep = Provider.autoDispose((ref) => 0);
-      final provider = AutoDisposeAsyncNotifierProvider<
-          AutoDisposeAsyncTestNotifier<int>, int>(
-        () => AutoDisposeAsyncTestNotifier((ref) {
-          return ref.watch(dep);
+      final provider = AutoDisposeStreamNotifierProvider<
+          AutoDisposeStreamTestNotifier<int>, int>(
+        () => AutoDisposeStreamTestNotifier((ref) {
+          return Stream.value(ref.watch(dep));
         }),
       );
       final container = createContainer();
+
+      // Skip the loading
+      await container.listen(provider.future, (previous, next) {}).read();
 
       expect(container.read(provider), const AsyncData(0));
     });
@@ -1017,8 +957,8 @@ void main() {
     // TODO use package:expect_error to test that commented lined are not compiling
 
     test('provider', () {
-      final provider = AsyncNotifierProvider<AsyncTestNotifier<int>, int>(
-        () => AsyncTestNotifier((ref) => 0),
+      final provider = StreamNotifierProvider<StreamTestNotifier<int>, int>(
+        () => StreamTestNotifier((ref) => Stream.value(0)),
       );
 
       provider.select((AsyncValue<int> value) => 0);
@@ -1034,20 +974,20 @@ void main() {
       canBeAssignedToRefreshable<Future<int>>(provider.future);
       canBeAssignedToAlwaysAliveRefreshable<Future<int>>(provider.future);
 
-      canBeAssignedToProviderListenable<AsyncNotifier<int>>(provider.notifier);
-      canBeAssignedToAlwaysAliveListenable<AsyncNotifier<int>>(
+      canBeAssignedToProviderListenable<StreamNotifier<int>>(provider.notifier);
+      canBeAssignedToAlwaysAliveListenable<StreamNotifier<int>>(
         provider.notifier,
       );
-      canBeAssignedToRefreshable<AsyncNotifier<int>>(provider.notifier);
-      canBeAssignedToAlwaysAliveRefreshable<AsyncNotifier<int>>(
+      canBeAssignedToRefreshable<StreamNotifier<int>>(provider.notifier);
+      canBeAssignedToAlwaysAliveRefreshable<StreamNotifier<int>>(
         provider.notifier,
       );
     });
 
     test('autoDispose', () {
-      final autoDispose = AsyncNotifierProvider.autoDispose<
-          AutoDisposeAsyncTestNotifier<int>, int>(
-        () => AutoDisposeAsyncTestNotifier((ref) => 0),
+      final autoDispose = StreamNotifierProvider.autoDispose<
+          AutoDisposeStreamTestNotifier<int>, int>(
+        () => AutoDisposeStreamTestNotifier((ref) => Stream.value(0)),
       );
 
       autoDispose.select((AsyncValue<int> value) => 0);
@@ -1063,24 +1003,24 @@ void main() {
       canBeAssignedToRefreshable<Future<int>>(autoDispose.future);
       // canBeAssignedToAlwaysAliveRefreshable<Future<int>>(autoDispose.future);
 
-      canBeAssignedToProviderListenable<AutoDisposeAsyncNotifier<int>>(
+      canBeAssignedToProviderListenable<AutoDisposeStreamNotifier<int>>(
         autoDispose.notifier,
       );
-      // canBeAssignedToAlwaysAliveListenable<AutoDisposeAsyncNotifier<int>>(
+      // canBeAssignedToAlwaysAliveListenable<AutoDisposeStreamNotifier<int>>(
       //   autoDispose.notifier,
       // );
-      canBeAssignedToRefreshable<AutoDisposeAsyncNotifier<int>>(
+      canBeAssignedToRefreshable<AutoDisposeStreamNotifier<int>>(
         autoDispose.notifier,
       );
-      // canBeAssignedToAlwaysAliveRefreshable<AutoDisposeAsyncNotifier<int>>(
+      // canBeAssignedToAlwaysAliveRefreshable<AutoDisposeStreamNotifier<int>>(
       //   autoDispose.notifier,
       // );
     });
 
     test('family', () {
-      final family = AsyncNotifierProvider.family<
-          AsyncTestNotifierFamily<String>, String, int>(
-        () => AsyncTestNotifierFamily((ref) => '0'),
+      final family = StreamNotifierProvider.family<
+          StreamTestNotifierFamily<String>, String, int>(
+        () => StreamTestNotifierFamily((ref) => Stream.value('0')),
       );
 
       family(0).select((AsyncValue<String> value) => 0);
@@ -1096,29 +1036,29 @@ void main() {
       canBeAssignedToRefreshable<Future<String>>(family(0).future);
       canBeAssignedToAlwaysAliveRefreshable<Future<String>>(family(0).future);
 
-      canBeAssignedToProviderListenable<FamilyAsyncNotifier<String, int>>(
+      canBeAssignedToProviderListenable<FamilyStreamNotifier<String, int>>(
         family(0).notifier,
       );
-      canBeAssignedToAlwaysAliveListenable<FamilyAsyncNotifier<String, int>>(
+      canBeAssignedToAlwaysAliveListenable<FamilyStreamNotifier<String, int>>(
         family(0).notifier,
       );
-      canBeAssignedToRefreshable<FamilyAsyncNotifier<String, int>>(
+      canBeAssignedToRefreshable<FamilyStreamNotifier<String, int>>(
         family(0).notifier,
       );
-      canBeAssignedToAlwaysAliveRefreshable<FamilyAsyncNotifier<String, int>>(
+      canBeAssignedToAlwaysAliveRefreshable<FamilyStreamNotifier<String, int>>(
         family(0).notifier,
       );
     });
 
     test('autoDisposeFamily', () {
       expect(
-        AsyncNotifierProvider.autoDispose.family,
-        same(AsyncNotifierProvider.family.autoDispose),
+        StreamNotifierProvider.autoDispose.family,
+        same(StreamNotifierProvider.family.autoDispose),
       );
 
-      final autoDisposeFamily = AsyncNotifierProvider.autoDispose
-          .family<AutoDisposeAsyncTestNotifierFamily<String>, String, int>(
-        () => AutoDisposeAsyncTestNotifierFamily((ref) => '0'),
+      final autoDisposeFamily = StreamNotifierProvider.autoDispose
+          .family<AutoDisposeStreamTestNotifierFamily<String>, String, int>(
+        () => AutoDisposeStreamTestNotifierFamily((ref) => Stream.value('0')),
       );
 
       autoDisposeFamily(0).select((AsyncValue<String> value) => 0);
@@ -1151,18 +1091,18 @@ void main() {
       // );
 
       canBeAssignedToProviderListenable<
-          AutoDisposeFamilyAsyncNotifier<String, int>>(
+          AutoDisposeFamilyStreamNotifier<String, int>>(
         autoDisposeFamily(0).notifier,
       );
       // canBeAssignedToAlwaysAliveListenable<
-      //     AutoDisposeFamilyAsyncNotifier<String, int>>(
+      //     AutoDisposeFamilyStreamNotifier<String, int>>(
       //   autoDisposeFamily(0).notifier,
       // );
-      canBeAssignedToRefreshable<AutoDisposeFamilyAsyncNotifier<String, int>>(
+      canBeAssignedToRefreshable<AutoDisposeFamilyStreamNotifier<String, int>>(
         autoDisposeFamily(0).notifier,
       );
       // canBeAssignedToAlwaysAliveRefreshable<
-      //     AutoDisposeFamilyAsyncNotifier<String, int>>(
+      //     AutoDisposeFamilyStreamNotifier<String, int>>(
       //   autoDisposeFamily(0).notifier,
       // );
     });
